+++ internal/config/types.go
@@ -40,8 +40,9 @@ const (
 type StorageType string
 
 const (
-	StorageLocal StorageType = "local"
-	StorageSMB   StorageType = "smb"
-	StorageNFS   StorageType = "nfs"
-	StorageSFTP  StorageType = "sftp"
+	StorageLocal  StorageType = "local"
+	StorageSMB    StorageType = "smb"
+	StorageNFS    StorageType = "nfs"
+	StorageSFTP   StorageType = "sftp"
+	StorageGDrive StorageType = "gdrive"
 )